+++ packages/common/src/core/schemas/internal/metrics/MetricSchema.ts
@@ -19,8 +19,11 @@ export const MetricSchema: IConfigurationSchema = {
     cardTitle: {
       type: "string",
     },
-    value: {
+    value: {},
+    valueType: {
       type: "string",
+      default: "string",
+      enum: ["string", "number", "date"],
     },
     dynamicMetric: {
       type: "object",
@@ -106,7 +109,10 @@ export const MetricSchema: IConfigurationSchema = {
       type: "string",
     },
   },
-  allOf: [{ $ref: "#/definitions/if-source-title-then-source-link" }],
+  allOf: [
+    { $ref: "#/definitions/if-source-title-then-source-link" },
+    { $ref: "#/definitions/value-type-value-mapping" },
+  ],
   definitions: {
     // TODO: reimplement popover with layouts release
     "if-layout-moreinfo-then-require-popover-title-description": {
@@ -125,6 +131,19 @@ export const MetricSchema: IConfigurationSchema = {
       },
       then: {
         required: ["sourceLink"],
+      },
+    },
+    "value-type-value-mapping": {
+      if: {
+        properties: {
+          valueType: { enum: ["string", "date"] },
+        },
+      },
+      then: {
+        properties: { value: { type: "string" } },
+      },
+      else: {
+        properties: { value: { type: "number" } },
       },
     },
   },