+++ packages/common/src/core/schemas/internal/metrics/ProjectUiSchemaMetrics.ts
@@ -5,6 +5,9 @@ import {
   SHOW_FOR_STATIC_RULE_ENTITY,
   SHOW_FOR_DYNAMIC_RULE_ENTITY,
   SHOW_FOR_SHARING_RULE_ENTITY,
+  SHOW_FOR_STATIC_AND_STRING_RULE_ENTITY,
+  SHOW_FOR_STATIC_AND_NUMBER_RULE_ENTITY,
+  SHOW_FOR_STATIC_AND_DATE_RULE_ENTITY,
 } from "./rules";
 
 /**
@@ -60,11 +63,64 @@ export const buildUiSchema = (
             },
           },
           {
+            scope: "/properties/_metric/properties/valueType",
+            type: "Control",
+            labelKey: `${i18nScope}.fields.metrics.valueType.label`,
+            rule: SHOW_FOR_STATIC_RULE_ENTITY,
+            options: {
+              control: "hub-field-input-tile-select",
+              layout: "horizontal",
+              helperText: {
+                labelKey: `${i18nScope}.fields.metrics.valueType.helperText`,
+                placement: "top",
+              },
+              enum: {
+                i18nScope: `${i18nScope}.fields.metrics.valueType.enum`,
+              },
+            },
+          },
+          {
             labelKey: `${i18nScope}.fields.metrics.value.label`,
             scope: "/properties/_metric/properties/value",
             type: "Control",
-            rule: SHOW_FOR_STATIC_RULE_ENTITY,
+            rule: SHOW_FOR_STATIC_AND_STRING_RULE_ENTITY,
             options: {
+              control: "hub-field-input-input",
+              messages: [
+                {
+                  type: "ERROR",
+                  keyword: "required",
+                  labelKey: `${i18nScope}.fields.metrics.value.message.required`,
+                  icon: true,
+                },
+              ],
+            },
+          },
+          {
+            labelKey: `${i18nScope}.fields.metrics.value.label`,
+            scope: "/properties/_metric/properties/value",
+            type: "Control",
+            rule: SHOW_FOR_STATIC_AND_NUMBER_RULE_ENTITY,
+            options: {
+              control: "hub-field-input-input",
+              type: "number",
+              messages: [
+                {
+                  type: "ERROR",
+                  keyword: "required",
+                  labelKey: `${i18nScope}.fields.metrics.value.message.required`,
+                  icon: true,
+                },
+              ],
+            },
+          },
+          {
+            labelKey: `${i18nScope}.fields.metrics.value.label`,
+            scope: "/properties/_metric/properties/value",
+            type: "Control",
+            rule: SHOW_FOR_STATIC_AND_DATE_RULE_ENTITY,
+            options: {
+              control: "hub-field-input-date",
               messages: [
                 {
                   type: "ERROR",
@@ -91,7 +147,7 @@ export const buildUiSchema = (
             options: {
               helperText: {
                 labelKey: `${i18nScope}.fields.metrics.unit.helperText`,
-                placement: "bottom",
+                placement: "top",
               },
             },
           },