+++ packages/common/src/projects/_internal/ProjectAssociationHierarchies.ts
@@ -1,0 +1,6 @@
+import { IHubAssociationHierarchy } from "../../associations/types";
+
+export const ProjectAssociationHierarchies: IHubAssociationHierarchy = {
+  children: [],
+  parents: ["initiative"],
+};