+++ packages/common/test/core/schemas/internal/metrics/ProjectUiSchemaMetrics.test.ts
@@ -47,9 +47,9 @@ describe("buildUiSchema: metric", () => {
               },
             },
             {
-              labelKey: `some.scope.fields.metrics.value.label`,
-              scope: "/properties/_metric/properties/value",
+              scope: "/properties/_metric/properties/valueType",
               type: "Control",
+              labelKey: `some.scope.fields.metrics.valueType.label`,
               rule: {
                 condition: {
                   scope: "/properties/_metric/properties/type",
@@ -58,6 +58,101 @@ describe("buildUiSchema: metric", () => {
                 effect: UiSchemaRuleEffects.SHOW,
               },
               options: {
+                control: "hub-field-input-tile-select",
+                layout: "horizontal",
+                helperText: {
+                  labelKey: `some.scope.fields.metrics.valueType.helperText`,
+                  placement: "top",
+                },
+                enum: {
+                  i18nScope: `some.scope.fields.metrics.valueType.enum`,
+                },
+              },
+            },
+            {
+              labelKey: `some.scope.fields.metrics.value.label`,
+              scope: "/properties/_metric/properties/value",
+              type: "Control",
+              rule: {
+                condition: {
+                  schema: {
+                    properties: {
+                      _metric: {
+                        properties: {
+                          type: { const: "static" },
+                          valueType: { const: "string" },
+                        },
+                      },
+                    },
+                  },
+                },
+                effect: UiSchemaRuleEffects.SHOW,
+              },
+              options: {
+                control: "hub-field-input-input",
+                messages: [
+                  {
+                    type: "ERROR",
+                    keyword: "required",
+                    labelKey: `some.scope.fields.metrics.value.message.required`,
+                    icon: true,
+                  },
+                ],
+              },
+            },
+            {
+              labelKey: `some.scope.fields.metrics.value.label`,
+              scope: "/properties/_metric/properties/value",
+              type: "Control",
+              rule: {
+                condition: {
+                  schema: {
+                    properties: {
+                      _metric: {
+                        properties: {
+                          type: { const: "static" },
+                          valueType: { const: "number" },
+                        },
+                      },
+                    },
+                  },
+                },
+                effect: UiSchemaRuleEffects.SHOW,
+              },
+              options: {
+                control: "hub-field-input-input",
+                type: "number",
+                messages: [
+                  {
+                    type: "ERROR",
+                    keyword: "required",
+                    labelKey: `some.scope.fields.metrics.value.message.required`,
+                    icon: true,
+                  },
+                ],
+              },
+            },
+            {
+              labelKey: `some.scope.fields.metrics.value.label`,
+              scope: "/properties/_metric/properties/value",
+              type: "Control",
+              rule: {
+                condition: {
+                  schema: {
+                    properties: {
+                      _metric: {
+                        properties: {
+                          type: { const: "static" },
+                          valueType: { const: "date" },
+                        },
+                      },
+                    },
+                  },
+                },
+                effect: UiSchemaRuleEffects.SHOW,
+              },
+              options: {
+                control: "hub-field-input-date",
                 messages: [
                   {
                     type: "ERROR",
@@ -90,7 +185,7 @@ describe("buildUiSchema: metric", () => {
               options: {
                 helperText: {
                   labelKey: `some.scope.fields.metrics.unit.helperText`,
-                  placement: "bottom",
+                  placement: "top",
                 },
               },
             },